#36 update projects

+++ src/test/java/ru/yandex/practicum/diplom2/api/ChangeUserCredentialsTest.java
@@ -148,8 +148,9 @@ public class ChangeUserCredentialsTest {
 
     @After
     public void clearDate() {
-        deleteUser(accessToken)
-                .statusCode(SC_ACCEPTED);
+        if(accessToken != null) {
+            deleteUser(accessToken);
+        }
     }
 
 }
